Support for DotNetCliTool references in CPS project
Resolves NuGet/Home#3845.

This change adds support for auto-restore of projects with tool references.
`VsSolutionRestoreService` is now capable of generating a `PackageSpec` for
each tool reference passed via the nomination.
As a key part of this change `ProjectSystemCache` has been updated to hold
`DependencyGraphSpec` object as an internal represenation of the project
restore info. All consuming project code has been changed to support the
new type.